backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-cloud-vision==3.5.0
pdfminer.six==20231228
pandas==2.2.0
numpy==1.26.3
python-multipart==0.0.7
aiofiles==23.2.1
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
pylint==3.0.3
mypy==1.8.0
black==23.12.1
isort==5.13.2
python-dotenv==1.0.0
